+++ HISTORY.rst
@@ -15,7 +15,7 @@ Features:
 - Add `missing` parameter to fields (:issue:`115`). Thanks :user:`philtay`.
 - Schema validators can take an optional ``raw_data`` argument which contains raw input data, incl. data not specified in the schema (:issue:`127`). Thanks :user:`ryanlowe0`.
 - Add ``validate.OneOf`` validator (:issue:`135`). Thanks :user:`philtay`.
-- Allow `NoneOf` error message to be interpolated with `{input}` and `{values}`.
+- Error messages for validators can be interpolated with `{input}` and other values (depending on the validator).
 
 Deprecation/Removals: