+++ HISTORY.rst
@@ -20,6 +20,7 @@ Features:
 Deprecation/Removals:
 
 - Remove support for generator functions that yield validators (:issue:`74`). Plain generators of validators are still supported.
+- The ``Select/Enum`` field is deprecated in favor of using `validate.OneOf` validator (:issue:`135`).
 
 1.2.2 (unreleased)
 ++++++++++++++++++